LibC:  - Implement tmpnam  - Implement (partially) tmpfile. This still needs to delete the file on close. We currently have no way of deleting a file other than through path. Which is not available at this point.  - Add system stub.

+++ librt/libc/include/stdlib.h
@@ -1,6 +1,7 @@
-/* MollenOS
+/**
+ * MollenOS
  *
- * Copyright 2011 - 2017, Philip Meulengracht
+ * Copyright 2017, Philip Meulengracht
  *
  * This program is free software : you can redistribute it and / or modify
  * it under the terms of the GNU General Public License as published by
@@ -27,7 +28,7 @@
 #include <locale.h>
 #include <wchar.h>
 
-#define EXIT_FAILURE    -1
+#define EXIT_FAILURE    (int)(-1)
 #define EXIT_SUCCESS    0
 
 #ifndef _SIZE_T_DEFINED
